modified

+++ src/STRIVER_SHEET/BINARYSEARCH/ma.java
@@ -2,7 +2,7 @@ package STRIVER_SHEET.BINARYSEARCH;
 
 public class ma {
     public static void main(String[] args) {
-        double pi = Math.PI;
+        long pi = (long) Math.PI;
         System.out.println(pi);
     }
 }
